Throttling OpenWeather API to one a second to comply with API limitations.
IMPORTANT: Removing dependency of the field from weather and weatherHourly tables.
Adding new table weather_station that will represent (for now) the result of the city id from OW API.
Adding dependency to field, weather and weatherHourly table to the new weather_station table.
This way we are clustering most of the OW API to the unique locations on the weather station table