Fix image and file access for Dockerized APIs
If the API is being run in a Docker container, image and file uploads to MinIO will need to use a Docker-accessible hostname -- either host.docker.internal or the Docker network internal name/service name -- in order for upload to complete. However, for the image or file to be later accessed/downloaded by the browser, the filename address must be localhost or your computer network name. As the previous funtionality was for the upload and download URL to be the same, this change adjusts the hostname at the point of download in the development environment. An optional .env variable allows moving off of the default of localhost. This issue only applies to those running the backend (and so uploading files and images) from within a Docker container.