LF-3421 Fixes Spanish translation
We need a single letter to define the day of the week. Tuesday
(Martes) and Wednesday (Miércoles) have the same starting letter
"M". The convention is to use a different letter for Miércoles "X".

Note: This is how Google Calendar translates the single letter day for
Miércoles.