LF-3468: Send notifications of the deleted tasks on crop plan delete
Problem: Users were supposed to receive a notification when a task that
they were assigned to is deleted. When a plan crop is deleted, all of
the crop tasks are also deleted and the users assigned to it do not
receive notifications.

The solution that was implemented: Use the function
taskController.sendTaskNotification() to send the notifications of the
deleted tasks in the managementPlanController.delManagementPlan()
function.