LF-3595 Pass revenue type to title of both forms; handle undefined first render
On the second (but not the first) visit to addSale, persistedFormData is initially undefined for one render, and then defined in the second render. This is the same behaviour that was seen in PR #2860. As in that PR, I'm addressing it here with optional chaining, but I'm not sure of the underlying reason for the undefined state."