LF-3360 Add "interested" validation to organic certifier survey
When requesting for a certification through API, the user is
able to get it even the farm is not pursuing it (interested = false).

This PR adds a validation to make sure the farm is interested on
certifications and returns 202 status code
[Accepted](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/204).
Also updates the test cases that are not mocking this property from
the DB therefore they were expecting a 200 status code.

Tests related to validate 202 can be added but it's a bit daunting the
amount of test cases and code in organicCertifierSurvey.test.js file
so I decided to skip that.