LF-4132 Remove from Price page the soft deleted sales
This PR modifies the raw SQL used to retrieve data regarding sales in
the Price page.
The constraint added to the query is "when sale is not deleted, display item".

The behaviour still the same as when there's no item in the chart
to be displayed.

Also activated some ignored tests cases that seems to be forgotten. Although more work is needed
in this particular test suite.