LF-4164 Group by field_array_id before updating details entries
This solves a bug that was a little edge-casey -- you would have to set up your basic cards and toggle create individual animal profiles in such a way that the new animals were inserted at the top of the list -- but that could potentially result in the previous solution's index-based updating to jumble data. This should be more robust.