LF-4771 Maintain the one to many relationship between notification and notification_users for IP notifications
Note: daily tasks due today has broken that relationship by calling the .notify() method in a loop over user_ids. This means that there is a one-to-one relationship between notifications and notification_user records for tasks due today; I don't know if that is intentional or not. The current flow will keep the pattern from weekly unassigned tasks and as suggested by the model